+++ plugins/nf-quilt/src/main/nextflow/quilt/QuiltObserver.groovy
@@ -16,14 +16,10 @@
 package nextflow.quilt
 
 import nextflow.quilt.jep.QuiltParser
-import nextflow.quilt.jep.QuiltPackage
 import nextflow.quilt.nio.QuiltPath
 import nextflow.quilt.nio.QuiltPathFactory
 
-import java.nio.file.Files
 import java.nio.file.Path
-import java.nio.file.Paths
-import java.time.LocalDateTime
 
 import groovy.transform.CompileStatic
 import groovy.util.logging.Slf4j
@@ -39,37 +35,8 @@ import nextflow.trace.TraceObserver
 @CompileStatic
 class QuiltObserver implements TraceObserver {
 
-    private final static String[] BIG_KEYS = [
-        'nextflow', 'commandLine', 'scriptFile', 'projectDir',
-        'homeDir', 'workDir', 'launchDir', 'manifest', 'configFiles'
-    ]
-    private final Set<QuiltPackage> pkgs = [] as Set
-    private Map config
-    private Map quiltConfig
+    private final Set<QuiltPath> paths = [] as Set
     private Session session
-
-    static void printMap(Map map, String title) {
-        log.info("\n\n\n# $title")
-        map.each {
-            key, value -> log.info("\n## ${key}: ${value}")
-        }
-    }
-
-    static void writeString(String text, QuiltPackage pkg, String filename) {
-        String dir = pkg.packageDest()
-        Path path  = Paths.get(dir, filename)
-        try {
-            Files.write(path, text.bytes)
-        }
-        catch (Exception e) {
-            log.error("writeString: cannot write `$text` to `$path` for `${pkg}`")
-        }
-    }
-
-    static String now() {
-        LocalDateTime time = LocalDateTime.now()
-        return time.toString()
-    }
 
     static QuiltPath asQuiltPath(Path path) {
         if (path in QuiltPath) {
@@ -101,9 +68,7 @@ class QuiltObserver implements TraceObserver {
     void onFlowCreate(Session session) {
         log.debug("`onFlowCreate` $this")
         this.session = session
-        this.config = session.config
-        this.quiltConfig = session.config.navigate('quilt') as Map
-        this.pkgs
+        this.paths
     }
 
     @Override
@@ -113,8 +78,8 @@ class QuiltObserver implements TraceObserver {
 
         if (qPath) {
             QuiltPath npath = normalizePath(qPath, session.getParams())
-            this.pkgs.add(npath.pkg())
-            log.debug("onFilePublish.QuiltPath[$qPath]: pkgs=${pkgs}")
+            this.paths.add(npath)
+            log.debug("onFilePublish.QuiltPath[$qPath]: paths=${paths}")
         } else {
             log.warn("onFilePublish.QuiltPath missing: $path")
         }
@@ -122,76 +87,9 @@ class QuiltObserver implements TraceObserver {
 
     @Override
     void onFlowComplete() {
-        log.debug("`onFlowComplete` ${pkgs}")
+        log.debug("`onFlowComplete` ${paths}")
         // publish pkgs to repository
-        this.pkgs.each { pkg -> publish(pkg) }
-    }
-
-    String readme(Map meta, String msg) {
-        return """
-            # ${now()}
-            ## $msg
-
-            ## workflow
-            ### scriptFile: ${meta['workflow']['scriptFile']}
-            ### sessionId: ${meta['workflow']['sessionId']}
-            - start: ${meta['time_start']}
-            - complete: ${meta['time_complete']}
-
-            ## processes
-            ${meta['workflow']['stats']['processes']}
-
-            """.stripIndent()
-    }
-
-    int publish(QuiltPackage pkg) {
-        String msg = pkg
-        Map meta = [pkg: msg]
-        String text = 'Stub README'
-        try {
-            meta = getMetadata()
-            msg = "${meta['config']['runName']}: ${meta['cmd']}"
-            text = readme(meta, msg)
-            meta.remove('config')
-        }
-        catch (Exception e) {
-            log.error('publish: cannot generate metadata (QuiltObserver uninitialized?)', e)
-        }
-        writeString(text, pkg, 'README.md')
-        writeString("$meta", pkg, 'quilt_metadata.txt')
-        def rc = pkg.push(msg, meta)
-        log.info("$rc: pushed package[$pkg] $msg")
-        if (rc > 0) {
-            print("ERROR[package push failed]: $pkg")
-        }
-
-        return rc
-    }
-
-    Map getMetadata() {
-        // TODO: Write out config files
-        Map cf = config
-        cf.remove('params')
-        cf.remove('session')
-        cf.remove('executor')
-        printMap(cf, 'config')
-        Map params = session.getParams()
-        params.remove('genomes')
-        params.remove('test_data')
-        printMap(params, 'params')
-        Map wf = session.getWorkflowMetadata().toMap()
-        String start = wf['start']
-        String complete = wf['complete']
-        String cmd = wf['commandLine']
-        BIG_KEYS.each { k -> wf[k] = "${wf[k]}" }
-        wf.remove('container')
-        wf.remove('start')
-        wf.remove('complete')
-        wf.remove('workflowStats')
-        wf.remove('commandLine')
-        printMap(wf, 'workflow')
-        log.info("\npublishing: ${wf['runName']}")
-        return [params: params, config: cf, workflow: wf, cmd: cmd, time_start: start, time_complete: complete]
+        this.paths.each { path -> new QuiltProduct(path, session) }
     }
 
 }